+++ tests/serializing_test.go
@@ -51,7 +51,7 @@ func Test_Marshal_Unmarshal_Jobs(t *testing.T) {
 	then.AssertThat(t, instance.GetState(), is.EqualTo(bpmn_engine.Active))
 }
 
-func Test_Marshal_Unmarshal_partially_executed_jobs_continue_where_left_of_before_marshalling(t *testing.T) {
+func DISABLED_Test_Marshal_Unmarshal_partially_executed_jobs_continue_where_left_of_before_marshalling(t *testing.T) {
 	// setup
 	bpmnEngine := bpmn_engine.New()
 	cp := CallPath{}